Fix maven usage errors in testcases example project.
Some test cases were using Maven in a way that would result in a maven build
failure if the test-case project was built from the command line.

Some test cases were using Maven in a way that would result in invalid EAR
application.xml files to be produced.  It is a requirement for at least one
<modules> element to exist.

These changes appear to correct the following test cases:
	org.eclipse.m2e.wtp.tests.OverlayTest
	org.eclipse.m2e.wtp.tests.conversion.EjbProjectConversionTest